+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Optional forced platform/version in ScenarioManager
 - DataManager.print_inputs_outputs_summary() method
 - inputs and outputs arguments to ScenarioManager.__init__
+- Mixed-case db_table_name warning in ScenarioDbTable
+- DB-table-name reserved word warning in ScenarioDbTable
 ### Removed
 - (minor) Removed support for DSX as platform choice